pyproject.toml
[project]
name = "weather-module"
version = "0.0"
description = ""
requires-python = ">=3.11"
dependencies = [
    "sekoia-automation-sdk>=1.0",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-cov",
    "requests",
    "requests-mock",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["weather_modules"]

[tool.black]
line-length = 119
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 119

[tool.coverage.run]
omit = [
    "tests/*",
    "main.py",
]

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [
    "tests",
]
addopts = '''
  --strict-markers
  --tb=short
  --cov=weather_modules
  --cov-branch
  --cache-clear
  --cov-report=html
  --cov-report=term-missing:skip-covered
  --cov-fail-under=5
  --capture=sys
'''
